chore: mockData 디렉토리 분리 (#48)
