+++ src/mocks/handlers/boardHandler.ts
@@ -1,11 +1,9 @@
 import { http, HttpResponse } from 'msw';
 
-import { boardList } from '../mockData';
-
-import { BoardDto } from '@/stores/repositories/BoardRepo';
+import { boardList } from './data/board';
 
 export const boardHandler = [
   http.get('/api/v1/boards', () => {
-    return HttpResponse.json<BoardDto[]>(boardList);
+    return HttpResponse.json<Board.ResponseDto[]>(boardList);
   }),
 ];
