+++ src/mocks/handlers/data/board.ts
@@ -1,0 +1,6 @@
+export const boardList: Board.ResponseDto[] = [
+  { id: '0', category: '공지게시판', name: '학생회 공지게시판' },
+  { id: '1', category: '공지게시판', name: '1학년 공지게시판' },
+  { id: '2', category: '자유게시판', name: '질문게시판' },
+  { id: '3', category: '자유게시판', name: '코딩게시판' },
+];
