+++ src/mocks/handlers/data/post.ts
@@ -1,0 +1,30 @@
+export const contentList: Post.Dto[] = [
+  {
+    id: '0',
+    title: 'post_title',
+    writerName: 'post_writerName',
+    writerAdmissionYear: 19,
+    writerProfileImage: 'content_writerProfileImage | null',
+    content: 'content',
+    createdAt: new Date('2024-01-11T15:34'),
+    updatedAt: new Date(),
+    numComment: 0,
+    updatable: true,
+    deletable: false,
+    isDeleted: false,
+  },
+  {
+    id: '1',
+    title: 'post_title2',
+    writerName: 'post_writerName2',
+    writerAdmissionYear: 20,
+    writerProfileImage: 'content_writerProfileImage2 | null',
+    content: 'content2',
+    createdAt: new Date('2024-01-12T15:34'),
+    updatedAt: new Date(),
+    numComment: 2,
+    updatable: false,
+    deletable: false,
+    isDeleted: false,
+  },
+];
