+++ src/mocks/handlers/data/replyComment.ts
@@ -1,0 +1,47 @@
+export const replyCommentList: ReplyComment.CreateResponseDto[] = [
+  {
+    postId: '0',
+    id: '0',
+    writerAdmissionYear: 19,
+    writerName: 'writer_name',
+    writerProfileImage: 'writerProfileImage | null',
+    content: 'replyCommentList_0_0',
+    createdAt: '2024-01-12T15:34',
+    updatedAt: '2024-01-13T15:34',
+    updatable: false,
+    deletable: false,
+    isDeleted: false,
+    refChildComment: 'refChildComment',
+    tagUserName: 'tagUserName',
+  },
+  {
+    postId: '0',
+    id: '1',
+    writerAdmissionYear: 19,
+    writerName: 'writer_name',
+    writerProfileImage: 'writerProfileImage | null',
+    content: 'replyCommentList_0_1',
+    createdAt: '2024-01-12T15:34',
+    updatedAt: '2024-01-13T15:34',
+    updatable: true,
+    deletable: false,
+    isDeleted: false,
+    refChildComment: 'refChildComment2',
+    tagUserName: 'tagUserName2',
+  },
+  {
+    postId: '0',
+    id: '2',
+    writerAdmissionYear: 19,
+    writerName: 'writer_name',
+    writerProfileImage: 'writerProfileImage | null',
+    content: 'replyCommentList_0_2',
+    createdAt: '2024-01-12T15:34',
+    updatedAt: '2024-01-13T15:34',
+    updatable: false,
+    deletable: true,
+    isDeleted: false,
+    refChildComment: 'refChildComment3',
+    tagUserName: 'tagUserName3',
+  },
+];
