+++ src/mocks/handlers/lockerHandler.ts
@@ -1,6 +1,5 @@
 import { HttpResponse, ResponseResolver, http } from 'msw';
-
-import { lockerAllLocationsList, lockerLocationsList } from '../mockData';
+import { lockerAllLocationsList, lockerLocationsList } from './data/locker';
 
 const getAllLocationHandler: ResponseResolver = () => {
   return HttpResponse.json<Locker.FindAllLocationResponseDto>(lockerAllLocationsList);
